refactor(lora): use int32_t parameters with per-value docs
Address review feedback on #565:

- LoraParameter values are now int32_t instead of float. A float can't
  hold a value like 906875000 Hz (906.875 MHz) exactly, so frequencies
  and rates are carried in base SI units (Hz, bit/s) and converted to
  RadioLib's MHz/kHz/kbps floats only at the doBegin() boundary.
- Move parameter documentation above each enum value so a missing
  docstring is visible at a glance.
- Document SYNC_WORD's value range (0x00-0xFF; conventionally 0x12 for
  private networks, 0x34 for public/LoRaWAN).
- Port the sx1262_radio.cpp adapter to int32_t to match the API.

+++ Drivers/sx126x-module/source/sx1262.cpp
@@ -291,13 +291,13 @@ static bool api_can_receive(Device* device, enum LoraModulation modulation) {
     return (radio != nullptr) && radio->canReceive(modulation);
 }
 
-static error_t api_set_parameter(Device* device, enum LoraParameter parameter, float value) {
+static error_t api_set_parameter(Device* device, enum LoraParameter parameter, int32_t value) {
     auto* radio = get_radio(device);
     if (radio == nullptr) return ERROR_INVALID_STATE;
     return radio->setParameter(parameter, value);
 }
 
-static error_t api_get_parameter(Device* device, enum LoraParameter parameter, float* value) {
+static error_t api_get_parameter(Device* device, enum LoraParameter parameter, int32_t* value) {
     auto* radio = get_radio(device);
     if (radio == nullptr) return ERROR_INVALID_STATE;
     return radio->getParameter(parameter, value);
